Final changes for make_uspex_flat to work.
